+++ inst/shinydashboard.js
@@ -43,16 +43,17 @@ var ensureActivatedTab = function() {
   // If there's a `data-start-selected` attribute and we can find a tab with
   // that name, activate it.
   var $startTab = $tablinks.filter("[data-start-selected='1']");
-  if ($startTab.length !== 0) {
-    $startTab.tab("show");
-    return;
+  if ($startTab.length === 0) {
+    // If no tab starts selected, use the first one, if present
+    $startTab = $tablinks.first();
   }
 
-  // If we got this far, just activate the first tab.
-  if (! $tablinks.parent("li").hasClass("active") ) {
-    $tablinks.first().tab("show");
+  // If there are no tabs, $startTab.length will be 0.
+  if ($startTab.length !== 0) {
+    $startTab.tab("show");
+
     $(".sidebarMenuSelectedTabItem").attr("data-value",
-      $tablinks.first().attr("data-value"));
+      $startTab.attr("data-value"));
   }
 };